Move apple track into own file
